+++ .nvmrc
@@ -1,0 +1,1 @@
+v10.15.3
