+++ lib/interface/cli/commands/gitops/install.cmd.js
@@ -2,7 +2,8 @@
 const _ = require('lodash');
 const Command = require('../../Command');
 const installRoot = require('../root/install.cmd');
-const { downloadProvider, detectProxy } = require('../hybrid/helper');
+const { detectProxy } = require('../../helpers/general');
+const { downloadProvider } = require('../hybrid/helper');
 const { Runner, components } = require('../../../../binary');
 
 const installArgoCmd = new Command({
