+++ lib/interface/cli/commands/hybrid/helper.js
@@ -950,36 +950,11 @@ function keyValueAsStringToObject(nodeSelectorStr) {
     }
 }
 
-function keyValueArrayToObject(arr) {
-    return arr.reduce((acc, cur) => {
-        const parts = cur.split('=');
-        // eslint-disable-next-line prefer-destructuring
-        acc[parts[0]] = parts[1];
-        return acc;
-    }, {});
-}
-
 function objectToKeyValueArray(obj) {
     return Object.keys(obj).reduce((acc, key) => {
         acc.push(`${key}=${obj[key]}`);
         return acc;
     }, []);
-}
-
-function addProxyVariables(envVars, { httpProxy, httpsProxy, noProxy }) {
-    const envVarsObj = keyValueArrayToObject(envVars);
-    if (httpsProxy && !envVarsObj.HTTPS_PROXY && !envVarsObj.https_proxy) {
-        envVars.push(`HTTPS_PROXY=${httpsProxy}`);
-        envVars.push(`https_proxy=${httpsProxy}`);
-    }
-    if (httpProxy && !envVarsObj.HTTP_PROXY && !envVarsObj.http_proxy) {
-        envVars.push(`HTTP_PROXY=${httpProxy}`);
-        envVars.push(`http_proxy=${httpProxy}`);
-    }
-    if (noProxy && !envVarsObj.NO_PROXY && !envVarsObj.no_proxy) {
-        envVars.push(`NO_PROXY=${noProxy}`);
-        envVars.push(`no_proxy=${noProxy}`);
-    }
 }
 
 function serealizeToKeyValuePairs(obj) {
@@ -989,17 +964,6 @@ function serealizeToKeyValuePairs(obj) {
         }
         return `${key}=${obj[key]}`;
     }, '');
-}
-
-function detectProxy() {
-    const httpProxy = process.env.http_proxy || process.env.HTTP_PROXY;
-    const httpsProxy = process.env.https_proxy || process.env.HTTPS_PROXY;
-    const noProxy = process.env.no_proxy || process.env.NO_PROXY;
-    return {
-        httpProxy,
-        httpsProxy,
-        noProxy,
-    };
 }
 
 function getRuntimeImagesWithRegistryUrl(registry) {
@@ -1088,7 +1052,6 @@ module.exports = {
     getRuntimeVersion,
     createTestPipeline,
     getTestPipeline,
-    addProxyVariables,
     updateTestPipelineRuntime,
     executeTestPipeline,
     createProgressBar,
@@ -1103,14 +1066,12 @@ module.exports = {
     newRuntimeName,
     newAgentName,
     keyValueAsStringToObject,
-    keyValueArrayToObject,
     objectToKeyValueArray,
     downloadRelatedComponents: downloadHybridComponents,
     downloadSteveDore,
     downloadVeonona,
     downloadProvider,
     runUpgrade,
-    detectProxy,
     serealizeToKeyValuePairs,
     getRuntimeImagesWithRegistryUrl,
     installAppProxy,
