+++ lib/interface/cli/commands/hybrid/init.cmd.js
@@ -4,6 +4,7 @@ const runnerRoot = require('../root/runner.cmd');
 const inquirer = require('inquirer');
 const colors = require('colors');
 const _ = require('lodash');
+const { addProxyVariables, detectProxy, keyValueArrayToObject } = require('../../helpers/general');
 const { getAllKubeContexts, getKubeContext } = require('../../helpers/kubernetes');
 const installMonitoring = require('../monitor/install.cmd');
 const createClusterCmd = require('../cluster/create.cmd');
@@ -16,7 +17,6 @@ const installationProgress = require('./installation-process');
 const { to } = require('./../../../../logic/cli-config/errors/awaitTo');
 const {
     createErrorHandler,
-    keyValueArrayToObject,
     createTestPipeline,
     executeTestPipeline,
     updateTestPipelineRuntime,
@@ -30,10 +30,8 @@ const {
     attachRuntime,
     newRuntimeName,
     newAgentName,
-    detectProxy,
     keyValueAsStringToObject,
     getRuntimeImagesWithRegistryUrl,
-    addProxyVariables,
     mergeValuesFromValuesFile,
     INSTALLATION_DEFAULTS,
 } = require('./helper');
