+++ lib/interface/cli/commands/pipeline/run.local.js
@@ -7,6 +7,7 @@ const _ = require('lodash');
 const { sdk } = require('../../../../logic');
 const { followLogs } = require('../../helpers/logs');
 const chalk = require('chalk');
+const { addProxyVariables, detectProxy } = require('../../helpers/general');
 
 const regex = /##[0-9a-f]{24}##/i;
 const EngineErrorPrefix = 'EngineError';
@@ -113,15 +114,18 @@ class RunLocalCommand extends RunBaseCommand {
         console.log(`Running pipeline: ${pipelineName}`);
 
         return new Promise((resolve, reject) => {
+            const proxyEnv = detectProxy();
+            const env = [
+                ...(debug ? ['PIPELINE_DEBUG_MODE=true'] : []),
+                `ACCESS_TOKEN=${currentContext.token}`,
+                `PIPELINE_ID=${pipelineName}`,
+                `BRANCH=${branch}`,
+                `CF_HOST=${currentContext.url}`,
+                'DOCKER_SOCKET_PATH=/var/run/docker.sock',
+            ];
+            addProxyVariables(env, proxyEnv);
             const eventEmitter = this.docker.run(DEFAULTS.ENGINE_IMAGE, [], undefined, _.mergeWith({
-                Env: [
-                    ...(debug ? ['PIPELINE_DEBUG_MODE=true'] : []),
-                    `ACCESS_TOKEN=${currentContext.token}`,
-                    `PIPELINE_ID=${pipelineName}`,
-                    `BRANCH=${branch}`,
-                    `CF_HOST=${currentContext.url}`,
-                    'DOCKER_SOCKET_PATH=/var/run/docker.sock',
-                ],
+                Env: env,
                 Hostconfig: {
                     Binds: [
                         '/var/run/docker.sock:/var/run/docker.sock',
@@ -131,8 +135,10 @@ class RunLocalCommand extends RunBaseCommand {
                 if (err) {
                     console.log(chalk.red(`Error when running pipeline : ${err}`));
                     // eslint-disable-next-line prefer-promise-reject-errors
-                    resolve(1);
+                    return resolve(1);
                 }
+
+                resolve(0);
             });
             if (this.argv.detach) {
                 resolve(0);
