+++ lib/interface/cli/helpers/general.js
@@ -182,6 +182,41 @@ function ignoreHttpError(e) {
     return undefined;
 }
 
+function detectProxy() {
+    const httpProxy = process.env.http_proxy || process.env.HTTP_PROXY;
+    const httpsProxy = process.env.https_proxy || process.env.HTTPS_PROXY;
+    const noProxy = process.env.no_proxy || process.env.NO_PROXY;
+    return {
+        httpProxy,
+        httpsProxy,
+        noProxy,
+    };
+}
+
+function keyValueArrayToObject(arr) {
+    return arr.reduce((acc, cur) => {
+        const parts = cur.split('=');
+        // eslint-disable-next-line prefer-destructuring
+        acc[parts[0]] = parts[1];
+        return acc;
+    }, {});
+}
+
+function addProxyVariables(envVars, { httpProxy, httpsProxy, noProxy }) {
+    const envVarsObj = keyValueArrayToObject(envVars);
+    if (httpsProxy && !envVarsObj.HTTPS_PROXY && !envVarsObj.https_proxy) {
+        envVars.push(`HTTPS_PROXY=${httpsProxy}`);
+        envVars.push(`https_proxy=${httpsProxy}`);
+    }
+    if (httpProxy && !envVarsObj.HTTP_PROXY && !envVarsObj.http_proxy) {
+        envVars.push(`HTTP_PROXY=${httpProxy}`);
+        envVars.push(`http_proxy=${httpProxy}`);
+    }
+    if (noProxy && !envVarsObj.NO_PROXY && !envVarsObj.no_proxy) {
+        envVars.push(`NO_PROXY=${noProxy}`);
+        envVars.push(`no_proxy=${noProxy}`);
+    }
+}
 
 module.exports = {
     wrapHandler,
@@ -194,4 +229,7 @@ module.exports = {
     watchFile,
     isDebug,
     ignoreHttpError,
+    detectProxy,
+    keyValueArrayToObject,
+    addProxyVariables,
 };
